fhir-mcp/pyproject.toml
[project]
name = "fhir-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.133.1",
    "fastmcp>=3.0.2",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.11.0",
]
